Handle empty DATA message exchange for EXTERNAL mechanism
go.dbus was responding to the server's initial advertisement
of EXTERNAL (SASL) auth:

    REJECT EXTERNAL

with an AUTH EXTERNAL containing an initial-response field:

    AUTH EXTERNAL [hex_encoded_username]

But if unix socket ownership based authentication is being used, the
D-Bus server does not expect any [initial-response] argument to AUTH
EXTERNAL. It responds with REJECTED EXTERNAL, i.e. a reject and a list
of allowed mechanisms.

Instead we should open the SASL handshake with an empty

    AUTH EXTERNAL

The server will respond with an empty DATA payload; we reply with our
own empty DATA payload, and the handshake is completed.

It's unclear under which circumstances the initial-data should be sent
or where it worked originally. It may not be correct to just never send
it; it might be required for some other SASL mechanisms. So it's not
totally clear that this fix won't break something else.

Really we should be using an actual SASL library to handle the payloads
of the DATA conversation to do SASL mechanism negotiation and
handshake, but that's far beyond the scope of this patch.

Issue #68